refactor: convert locations and shift types to server components
- Replace /api/locations and /api/shift-types API routes with direct server-side data fetching
- Convert registration and profile edit pages from client components to server component + client wrapper pattern
- Fetch data directly from Prisma in server components using Promise.all for parallel queries
- Remove API routes from middleware public routes list
- Update authentication documentation to remove references to deleted API routes

Benefits:
- Eliminates unnecessary API roundtrips for static/semi-static data
- Better performance with server-side data fetching
- Reduced JavaScript bundle size
- More consistent with Next.js App Router best practices
